backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
python-multipart
pydantic-settings
